+++ include/view/error.h
@@ -1,9 +1,0 @@
-#ifndef VIEW_ERROR_H
-#define VIEW_ERROR_H
-
-#include <Python.h>
-#include <view/backport.h>
-
-extern PyTypeObject ErrorType;
-
-#endif