+++ include/view/view.h
@@ -7,7 +7,6 @@
 #include <view/map.h>
 #include <view/context.h>
 #include <view/tcpublic.h>
-#include <view/error.h>
 
 void view_fatal(
     const char* message,
@@ -17,6 +16,7 @@ void view_fatal(
 );
 
 extern PyObject* ip_address;
+extern PyObject* invalid_status_error;
 
 #if defined(__LINE__) && defined(__FILE__)
 #define VIEW_FATAL(msg) view_fatal(msg, __FILE__, __func__, __LINE__)