+++ CHANGELOG.md
@@ -23,9 +23,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added the `view build` command
 - Added `App.template`
 - Route errors now display the error message when `dev` is `True`
-- Changed exception rendering in route errors
+- Changed exception rendering in route errors to use the `rich` renderer
 - Added `compile_type` and `TCValidator`
 - Added `markdown` and `App.markdown`
+- Added the `Error` class
+- Added the `error_class` parameter to both `new_app` and `App`
 - **Breaking Change:** The `body` parameter in `Response` is now required
 
 ## [1.0.0-alpha8] - 2024-1-21